requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
pip-chill==1.0.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
